+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 ### Added
 - Calendar Widget to visualize repetitions left for months/days (heatmap coloration, plugged to backend)
 
+### Fixed
+- In node tree, left triangle was grayed out even if the current node had a non-dismissed fragment child, when that fragment only had spore children. It is not the case anymore.
+
+
 ## v0.0.4
 ### Added
 - Warning message when undoing a review of a deleted node.